quota: Avoid double callbacks in QuotaManagerImpl::BucketDataDeleter.
Before this CL, BucketDataDeleter's constructor accepts two callbacks,
`callback` and `completion_closure`, and runs them in order. This adds
complexity to BucketDataDeleter's API contract (and implementation). For
example, the current implementation assumes that the `callback` callback
does not destroy the BucketDataDeleter instance.

This CL removes the extra complexity by having BucketDataDeleter accept
a single callback. Fortunately, this adds minimal complexity to the code
using BucketDataDeleter.

This CL also adds an unit test covering the behavior of
QuotaManagerImpl::DeleteBucketData() when the callback deletes the
QuotaManagerImpl. This test would fail before the simplification
introduced in this CL, because QuotaManagerImpl::DeleteBucketData()'s
callback ended up being passed as the `callback` callback to
BucketDataDeleter. When `callback` deleted QuotaManagerImpl, it
indirectly deleted the BucketDataDeleter. This broke the
BucketDataDeleter implementation's assumption that `callback` would not
delete BucketDataDeleter.

The newly introduced unit test is a reduced reproduction of crashers
encountered in the field.
StoragePartitionImpl::QuotaManagedDataDeletionHelper::ClearBucketsOnIOThread()
calls QuotaManager::DeleteBucketData() with a callback that points to
the PerformQuotaManagerStorageCleanup() function and holds a reference
to QuotaManager. During browser shutdown, this QuotaManager reference
may be the only thing keeping the QuotaManagerImpl alive. In this case,
running the callback deletes QuotaManagerImpl, matching the scenario
covered by the new unit test.

Bug: 1292216
Change-Id: I672ada38f955f54475553400486038b55846a9a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3460563
Auto-Submit: Victor Costan <pwnall@chromium.org>
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Ayu Ishii <ayui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970722}