Add ConvertImage to ColorConversionSkFilterCache
This function will (when needed) use the filter cache to convert an
SkImage to another color space.

This will be used by the image decode caches in cc/paint, which
cannot depend on ui/gfx. Move ColorConversionSkFilterCache and
ColorTransform to ui/gfx:color_space to avoid a circular dependency.

Plumb it to ServiceImageTransferCacheEntry. Gate its behavior on a
base::Feature (left disabled), and leave the parameters as stand-in
values (for now).

Bug: 1286088
Change-Id: I0afae5b3591e367896c14285e4e1dacabe264b98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3418521
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970723}