quota: Avoid double callbacks in QuotaManagerImpl::HostDataDeleter.
Before this CL, HostDataDeleter's constructor accepts two callbacks,
`callback` and `completion_closure`, and runs them in order. This adds
complexity to HostDataDeleter's API contract (and implementation). For
example, the current implementation assumes that the `callback` callback
does not destroy the BucketDataDeleter instance.

This CL removes the extra complexity by having HostDataDeleter accept
a single callback. Fortunately, this adds minimal complexity to the code
using HostDataDeleter.

This CL also adds an unit test covering the behavior of
QuotaManagerImpl::DeleteHostData() when the callback deletes the
QuotaManagerImpl. This test would fail before the simplification
introduced in this CL, because QuotaManagerImpl::DeleteHostData()'s
callback ended up being passed as the `callback` callback to
HostDataDeleter. When `callback` deleted QuotaManagerImpl, it
indirectly deleted the HostDataDeleter. This broke the
HostDataDeleter implementation's assumption that `callback` would not
delete HostDataDeleter.

The newly introduced unit test covers a scenario that is hypothetically
possible (and therefore should be handled), but doesn't currently occur
in Chromium.

Change-Id: I74f5ff0b054acefdd5ff21d47aeb9c8acbf91788
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3460339
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970766}