fido/mac: Use SecTask instead of blocking and broken SecCode
device::fido::mac::TouchIdContext wants to determine whether the main
executable is signed with the entitlement required for Touch ID to work
before advertising this capability to the web. Prior to this change, it
did so by examining the main executable’s entitlements via the SecCode
interface. SecCode (specifically SecStaticCode, but you can’t avoid
SecCode’s use of SecStaticCode) pokes around on the disk. This means
that it’s potentially blocking, and there’s a disassociation between the
effective entitlements (set at launch time) and the entitlements found
on disk (because the main executable’s path may no longer correspond to
the vnode associated with the running executable). Where an application
has been updated on disk but not yet relaunched, the file may be
completely different, and other bugs in SecCode may result in it
erroneously determining that no entitlements are present at all, even
when both the old and new files on disk were properly entitled.

Instead of SecCode, SecTask will now be used. Internally, SecTask
consults the live copy of code signing attributes in the kernel, without
touching the disk at all. SecTask’s view of the process’ entitlements is
thus consistent with the process’ effective entitlements, and can be
obtained without blocking.

This preserves the application’s ability to use Touch ID even when an
update has been applied but the application has not yet been relaunched,
while also preserving the application’s ability to check whether it has
been signed with the appropriate entitlement to access this feature.

Bug: 1295403, 1296607
Change-Id: I69b6f5a23627c4c057ebab46534aff6a68a8ec50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3462051
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Mark Mentovai <mark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970773}