[Launcher Sort] Add tests to verify the fade in animation interruption
This CL ensures that interrupting fade in animation is tested.
Now only the fade out animation interruption is tested. To verify
fade in animation interruption, this CL does the following things:

(1) Allow the test class to install the callback that runs at the end
of the fade out animation.

(2) The reorder animation done callback for testing carries an
additional parameter to specify the animation type.

In addition, this CL aborts the reorder animation on the paged apps
grid when the tablet mode changes.

Bug: 1294100
Change-Id: Iddd26068972b4c502da7c39e96b668331eb0fb8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3438828
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970787}