ash: Fix ToastManager queueing multiple existing toasts.
Fix multiple toasts of the same toast id being queued when
ToastManagerImpl::Show() is called before the current toast has finished
hiding. This was because previously, if a call to Show() was made for
a toast with an id that was the same as the current toast's id, the
new toast would be added to the front of the queue and the current toast
hidden. Once the current toast's hiding animation has finished, it would
pop the replacement toast from the top of the queue. However, if another
call to Show() was made with the same id before the hiding animation
finished, then a second toast would also be added to the front of the
queue. Therefore, once the current toast finishes hiding, the most
recent toast would (correctly) be shown, but once that toast has
finished, the first, obsolete toast would be shown. This CL fixes this
by rearranging the logic in Show() to check for existing toasts with
the same id in the queue first, and if one exists then that toast is
replaced, so that the second replacement toast would replace the first
replacement toast.

Video:
https://drive.google.com/file/d/1CFwgr5yRY6yO58LNpTkfu6h38J7O3THB/view?usp=sharing

Fixed: b/217777512
Change-Id: I817cb3075fc570edd41888e74a5445c2951ff842
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3452664
Auto-Submit: Gordon Seto <gordonseto@google.com>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970789}