[Shortcuts] Set details text of KeyboardShortcutResult.
Change summary:

(1) Add a new method CreateTextVectorFromTemplateString() to the
KeyboardShortcutResult class. This is used to parse a shortcut template
string (e.g. "Press $1$2$3") into different types of TextItem
elements, which are returned in a TextVector.

(2) Use the method in (1) to set the details text of the
KeyboardShortcutResult / ChromeSearchResult.

(3) To support (1), add a new method GetIconCodeFromKeyboardCode() which
maps:

  ui::KeyboardCode codes (used by the backend) to
  ash::SearchResultTextItem::IconCode (used by the frontend).

If a KeyboardCode is unsupported by the frontend, returns absl::nullopt.

Test: unit tests
Bug: 1290682
Change-Id: I35c8f182da91ddd42226bff2f5e64a187e193390
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3459448
Reviewed-by: Tony Yeoman <tby@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Amanda Deacon <amandadeacon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970847}