[Intents] Remove No-Handler external protocol dialog for non Tel-links
Previously, when the user clicks on a URL that the device doesn't have
a handler for, a dialog would appear. This dialog was unhelpful and
didn’t match the match behaviour of Windows/Linux/Mac devices (on
these other devices, no dialog appears and nothing happens). This CL
makes the change to align CrOS behaviour with Windows/Linux/Mac for
non Tel-link schemes. For Tel-link schemes, we still show the relevant
no-handler dialog.

Change-Id: I29be64a2ae7b6bef6cea49745c43d9522183ecd8
Bug: 1218108
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3452906
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Vivian Pao <vpao@google.com>
Cr-Commit-Position: refs/heads/main@{#970855}