WebAppShimManagerDelegate: Fix AppCanCreateHost and AppUsesRemoteCocoa
These functions should return true only for PWAs or bookmark apps that
open in windows.

For apps that do not open in windows, returning true resulted in an
AppShimManager::ProfileState being created. A ProfileState is a
structure that exists to maintain a record of the Browsers for a
Profile, for an app. The ProfileState is deleted when all Browsers
are removed (and therefore the ProfileState cannot outlive the
Profile).

But, because this zombie ProfileState was created with no Browsers,
it was also never deleted, and kept a raw pointer to the Profile,
which eventually was poked and caused crashes.

This issue only appeared now because we switched from using
ExtensionAppShimManagerDelegate (which had the correct behavior)
to WebAppShimManagerDelegate (which didn't). A more robust fix
would be to more tightly couple the existence of a ProfileState to
the presence of a Browser instance (in this path, it is created in
anticipation of the creation of a Browser instance).

Bug: 1278260
Change-Id: I8920d5d6e01e0c905433c0f48a35a0183055fe8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3452669
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970857}