Deflake and re-enable MediaAppUIWithAudioGtestBrowserTest.SingleAudioLaunch
This was flaky in the "mock" app we use on open source bots because its
error handling is not as sophisticated as the real app. The real app has
a `console.warn` on this codepath, but the mock will `error`.

The flake occurs because sometimes a call to `play()` on a zero-byte
Blob happens before the test ends. That throws an exception that the
mock app sends to console.error.

We do still want the error diagnostic for integration tests that are
using "real" audio files, so instead detect in the mock if a zero-byte
file is being loaded and bail. The real media_app will continue to
console.warn in its error handling, but that won't fail the test.

Fixed: b/217439041
Change-Id: Iccd2d6d1042662c26c37fd0b3642c03b01b9d98e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3459373
Reviewed-by: Mitch McDermott <mcdermottm@google.com>
Commit-Queue: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970862}