[CrOS Bluetooth] Show battery notification when device is connected
When a device is connected, and the device has a low battery, a low
battery notification is not shown, only when the battery level updates
does a notification get shown. This CL fixes that issue and notifies
observers of a device connection.

Fixed: b/217940117
Change-Id: I5495d694b41b325537a866e7bfa2e6633c23ecdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3450461
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Theo Johnson-kanu <tjohnsonkanu@google.com>
Cr-Commit-Position: refs/heads/main@{#970874}