Use today's circle for focusing
The selected date background the same as the focusing background.
As discussed with our UX team, will change the focusing circle to the
same as today's circle for now.

Also fix a row height calculation for date cell focusing. We should
add the paddings.

Bug: 1295358
Change-Id: I17c4d2f2b7e1353d96283cf88c2be201e9c568cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3453289
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Jiaming Cheng <jiamingc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970877}