Speculative fix for crash in ReleaseColorSwapchainImages
Adds a nullptr check for SharedImageInterface to speculatively fix a
crash. If this does not resolve the issue, then there is likely an
issue in the context_provider_ implementation, given that we hold the
context_provider as a scoped_refptr and its validity is already checked.

Fixed: 1295428
Change-Id: I67c055b44063c9fbcfae459353dc59424f477c44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3462056
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970917}