WebXR: add traces to session support/creation requests
Also make the error strings unique so that it's easier to figure out the
exact reason why the session support / creation failed.

With this change, a trace paired with the console output should be
sufficient to have an approximate idea about which code path caused an
issue with session creation.

Change-Id: Iaee19cdfd95c684ad6087eb38c1b092c2244771d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2998796
Commit-Queue: Piotr Bialecki <bialpio@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970920}