[exo] Fix loss of buffer attachment when setting rounded corners.
Additionally, clarify how to use pending/cached/set state in
exo::Surface. This fixes videos flashing on scroll that are also in
hardware overlay in lacros.

Setting rounded corner bounds sets has_pending_contents_, which was
previously only set on attach. When the buffer is moved, it becomes
null so if rounded corner bounds are set without a new buffer being
attached, the existing buffer is cleared. Scrolling happens to change
the position of the rounded corner bounds faster than attaching new
buffers which explains the issue.

It's unfortunate for this CL that the same struct is used for pending,
cached, and actual state - this means that the actual state which
logically shouldn't be nullopt now has a tristate of being not defined
(null option) that should never actually occur. It is necessary to
distinguish the null option from a set but empty value in the pending
and cached states, however (e.g. buffer).

Bug: 1283305
Test: scrolling on video web page with lacros doesn't cause flashing
Change-Id: Iba021c59a89a7be8e838241d95bd05f690d5ca50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3448043
Reviewed-by: Kramer Ge <fangzhoug@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970931}