[journeys] Move post-processing of clusters into `QueryClustersState`
This CL moves the post-processing step of clusters into
`QueryClustersState`. This allows us to add more post-processing steps
in a clean fashion later.

This also simplifies the Service and a service test.

It adds a corresponding test to the `QueryClustersState` to keep
coverage the same.

Bug: 1292134
Change-Id: I069a469a030e0867111fa08922633a237e159fc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3456745
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970933}