Add in branding strings and build flag for Reven
We need to update and add in new strings for the upcoming reven product
launch.

This CL adds in a new GN arg called is_reven that is used alongside all
the chromeos conditionals for determining the desired text. The arg
defaults to false, and should be set to true for the reven board and
CloudReady 2.0 builds.

Note that we are not requiring string translations in this CL. This is
because this CL is mainly focused around implementing the logic to
differentiate the new brand, and will be immediately updated with new
strings after launch.

Bug: b:189941268
Change-Id: I36f875ac114a2c868199595b059d609e87fe4df0
Skip-Translation-Screenshots-Check: True
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3455027
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Richard Chui <richui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970958}