[journeys] Add a state object that lives on the post-processing sequence
We have `QueryClustersState` which is designed to hold the pagination
state of the Journeys UI.

Now we want to do some fancier post-processing things, such as
deduplicating single-visit clusters across batches.

This requires us to hold some state (like a set of seen URLs) that can
be accessed on the post-processing sequence.

This CL adds the `QueryClustersState::PostProcessor` inner object that
contains that state, and manages its lifetime. This CL doesn't actually
do anything with that yet, and so should have no user-visible effects.

A followup CL will actually use this new `PostProcessor` object.

Bug: 1292134
Change-Id: If35d0b4021160356731c030c1a74ef8218d9d767
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3462447
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971006}