Prerender: Stop location bar from displaying pf=cs
This parameter is set for prerendering navigation, and after the
prerendered page is activated, the location bar should not display this
parameter, so that the prerendered page looks like a non-prerendered
page for users.

The CL makes use of the VirtualUrl, which allows caller to set a URL
that differs from the loaded url.
This approach achieves the following goals:
- Do not display pf=cs on the location bar.
- BFCache can use the cached page(fetched for url with pf=cs).
- Back/Forward navigation can reuse the existing HTTP cache.
- If one reloads a page by clicking 'enter', this loads the correct url.
  (without pf=cs)
This approach does not achieve the following goal:
- If one reloads by clicking the reload button, this loads the wrong
  url (with pf=cs)

Bug: 1291147
Change-Id: I561bab9c037e52e991967ed9d6e6abf5b8102459
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3450068
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971007}