[Launcher Sort] Generalize the methods to calculate position in order
Currently, there are methods such as `CalculateNewItemPosition()` to
calculate the position that maintains the app list sort order for a
new item. This CL generalizes these methods so that they can handle
the items existing in the model updater as well.

There is one difference between handling a new item and handling an
existing item: for an existing item, if its current position already
maintains the app list order, its position should not change. This CL
implements this feature and add unit tests to protect it.

Bug: 1293162
Change-Id: I47576e5084cf934c97ec45029160c82a4cac007c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3462553
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971025}