[lacros] Fix Session Restore for Lacros Primary
This CL fixes session restore behavior for lacros when lacros-chrome
windows are opened and lacros-chrome is already running in windowless
mode.

When a new lacros-chrome window is opened and the process is already
running, BrowserLacrosService calls through to the NewEmptyWindow()
API, which calls SessionService::RestoreIfNecessary() to handle
session restore when `should_trigger_session_restore` is true.

Currently SessionService::RestoreIfNecessary() has special handling
for ash-chrome that handles the case where session restore is
requested but there has been a previous unclean exit (launching an
empty browser and showing the crash recovery bubble).

This CL enables this special handling for lacros-chrome to allow it
to handle the case where chrome is already running but the first
window must show a new window with the crash recovery bubble.

Currently the kNoStartupWindow switch is set when lacros-chrome is
launched with InitialBrowserAction::kDoNotOpenWindow. The intention
of this is to prevent lacros-chrome from launching a window during
startup. However this switch remains set throughout the entire
execution of chrome.

This leads to issues since SessionService uses LaunchBrowser() here
to create the first Browser window when the Browser icon is clicked
on the shelf. In this case kNoStartupWindow will prevent the Browser
window from ever launching. As a temporary workaround this CL removes
the kNoStartupWindow switch from the command line when launching the
Browser window in RestoreIfNecessary().

HasPendingUncleanExit() has been updated to check if a browser window
has been opened for a given profile rather than checking whether the
profile has been launched. This has been done as a lacros-chrome
process started in a background state will be recorded as launched but
the correct but the unclean exit will only have been handled after a
browser window has been opened.

Bug: 1286895
Change-Id: Ib89837600e033c944d395fede1b9bd27733406ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3445271
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Jan Krčál <jkrcal@chromium.org>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971032}