Don't respond to queries until FirstPartySets is completely initialized.
This CL makes Chromium wait until the FirstPartySets instance is fully
initialized before doing anything that depends on the set of First-Party
Sets that are available (e.g. issuing a network request that involves
cookies). The instance is considered fully initialized if the feature is
disabled, or the instance has received and parsed all of the required
inputs.

Note that none of the required inputs involve making a network request,
so this will not deadlock. Also note that the FirstPartySets instance
always receives all of the inputs it requires, even if component updater
is disabled, or no component exists on disk, or the component could not
be read properly.

This CL also removes the ability to "synchronously" inspect internal
First-Party Sets state (i.e. getting the size of the map) from tests,
since it's no longer necessary to poll until FirstPartySets has been
re-initialized.

Bug: 1283271, 1266014
Change-Id: Ic90aefc7c46c21a2c347e2afef980080078b25dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3449280
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Chris Fredrickson <cfredric@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971043}