Set blurred semi-transparent background to popup notifications
This CL sets the blurred semi-transparent background to popup ARC
notifications to keep the UX consistency with ash notifications.
With the nt-refresh flag enabled, ARC notification content has a
transparent background so that it inherits the background of the
notification center, but when it’s a popup notification we need to set
the background because the notification “floats” on the desktop.

BUG=b:218607170

Change-Id: Ibb511dbf79a22b2d469b8c4f5c9b8670c8862c73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3459375
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971062}