Add UseDefaultSettings to PrintBackendServiceManager
If the print backend process terminates then it is necessary to have
extra handling to ensure that callbacks are still made, as otherwise the
UI can end up in a stuck state.

Extend PrintBackendServiceManager to include a wrapper around the
UseDefaultSettings() call, to include protection against this scenario.

Bug: 809738
Change-Id: Ic1a16a3f21ba3eb450eb906e97ee5f64546e8b15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3313947
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971068}